beta1
框架完成，jpa读取mysql

+++ mirror-webapp/src/main/java/com/mirror/service/TestService.java
@@ -3,5 +3,5 @@ package com.mirror.service;
 
 public interface TestService {
 
-	public String getUserByID(int id);
+	public String getUserByID(Long id);
 }
